Removed useless components. Using unique dashboard.
