+++ src/components/App.jsx
@@ -1,12 +1,20 @@
 import React from 'react';
 import Auth0Lock from 'auth0-lock';
+import $ from 'jquery';
 
 import config from 'config';
 
 import BaseComponent from './common/BaseComponent';
 import Home from './Home';
-import UserDashboard from './UserDashboard';
-import UserListDashboard from './UserListDashboard';
+import Dashboard from './Dashboard';
+
+const geolocation = (
+  navigator.geolocation || {
+    getCurrentPosition: (success, failure) => {
+      failure('Your browser doesn\'t support geolocation.');
+    },
+  }
+);
 
 class App extends BaseComponent {
 
@@ -27,6 +35,102 @@ class App extends BaseComponent {
     this.setState({ idToken: this.getIdToken() });
   }
 
+  registerUserLocation(user) {
+    const updatedLocation = {
+      latitude: user.latitude,
+      longitude: user.longitude,
+    };
+    $.ajax({
+      url: `/api/users/${user.user_id}`,
+      contentType: 'application/json',
+      headers: { 'If-Match': user._etag },
+      type: 'PATCH',
+      data: JSON.stringify(updatedLocation),
+      success: function registerUserLocationSuccess() {
+        console.log('User position updated.');
+      },
+      error: function registerUserLocationError(xhr, status, err) {
+        console.error(this.props.url, status, err.toString());
+      }.bind(this),
+    });
+  }
+
+  registerNewUser(profile) {
+    const user = {
+      user_id: profile.user_id,
+      firstname: profile.given_name,
+      lastname: profile.family_name,
+      email: '',
+      city: profile.location.name,
+      profession: profile.headline,
+      latitude: profile.latitude,
+      longitude: profile.longitude,
+    };
+    $.ajax({
+      url: '/api/users',
+      contentType: 'application/json',
+      type: 'POST',
+      data: JSON.stringify(user),
+      success: function registerNewUserSuccess() {
+        console.log('User created');
+        this.setState({ profile: user });
+      }.bind(this),
+      error: function registerNewUserError(xhr, status, err) {
+        console.error(this.props.url, status, err.toString());
+      }.bind(this),
+    });
+  }
+
+  getUserLocationAndExecCallback(userProfile, callback) {
+    const user = userProfile;
+    user.latitude = 0.0;
+    user.longitude = 0.0;
+    geolocation.getCurrentPosition((position) => {
+      user.latitude = position.coords.latitude;
+      user.longitude = position.coords.longitude;
+      callback(user);
+    }, (reason) => {
+      console.log(`Unable to retrieve current position: ${reason}`);
+      callback(user);
+    });
+  }
+
+  registerUserIfNotExisting(profile) {
+    $.ajax({
+      url: '/api/users',
+      data: {
+        user_id: profile.user_id,
+      },
+      dataType: 'json',
+      cache: false,
+      success: function registerUserIfNotExistingSuccess(data) {
+        if (data._items.length !== 0) {
+          console.log('User exists!');
+          const userProfile = data._items[0];
+          this.setState({ profile: userProfile });
+          this.getUserLocationAndExecCallback(userProfile, this.registerUserLocation.bind(this));
+        } else {
+          console.log('User do not exists! It will be created.');
+          const userProfile = profile;
+          this.getUserLocationAndExecCallback(userProfile, this.registerNewUser.bind(this));
+        }
+      }.bind(this),
+      error: function registerUserIfNotExistingError(xhr, status, err) {
+        console.error(this.props.url, status, err.toString());
+      }.bind(this),
+    });
+  }
+
+  getProfileAndUpdateUser(idToken) {
+    this.lock.getProfile(idToken, (err, profile) => {
+      if (err) {
+        console.log('Error loading the Profile', err);
+        return;
+      }
+      this.registerUserIfNotExisting(profile);
+    });
+  }
+
   getIdToken() {
     let idToken = localStorage.getItem('userToken');
     const authHash = this.lock.parseHash(window.location.hash);
@@ -34,6 +138,7 @@ class App extends BaseComponent {
       if (authHash.id_token) {
         idToken = authHash.id_token;
         localStorage.setItem('userToken', authHash.id_token);
+        this.getProfileAndUpdateUser(idToken);
       }
       if (authHash.error) {
         console.log('Error signing in', authHash);
@@ -44,24 +149,10 @@ class App extends BaseComponent {
   }
 
   render() {
-    let dashboard = null;
-    switch (this.props.display) {
-      case 'userList':
-        dashboard = (
-          <UserListDashboard />
-        );
-        break;
-      default:
-        dashboard = (
-          <UserDashboard lock={this.lock} idToken={this.state.idToken} />
-        );
-        break;
-    }
-
     if (this.state.idToken) {
       return (
         <div>
-        { dashboard }
+          <Dashboard />
         </div>
       );
     }
