--- src/components/BottomUserPanel.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import UserWorldMap from './UserWorldMap';
-
-const BottomUserPanel = function BottomUserPanel() {
-  return (<UserWorldMap />);
-};
-
-export default BottomUserPanel;
