+++ src/components/MapPanel.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import WorldMap from './WorldMap';
+
+const MapPanel = function MapPanel(props) {
+  return (<WorldMap users={props.users} />);
+};
+
+MapPanel.propTypes = {
+  users: React.PropTypes.array,
+};
+
+export default MapPanel;
